整理文章管理接口
