+++ ds/blog/src/main/java/com/mylog/ds/blog/entity/dto/ArticleDto.java
@@ -36,4 +36,6 @@ public class ArticleDto {
      */
     private String isLock;
 
+    private String sendPlace;
+
 }
